Allow to open buffers in vertical splits
Previously, when using `vert Jqplsit`, the filter buffer would occupy
the whole screen at the bottom, which was unexpected to me:

 -------- --------
|        |        |
|        |        |
|        |        |
|-----------------|
|                 |
 -------- --------

The layout I expected would be something like this:

 -------- --------
|        |        |
|        |        |
|        |--------|
|        |        |
|        |        |
 -------- --------

Instead of using the 'botright' modifier to open the filter buffer, I
check if the output buffer was opened with a vertical split or not.

If it was opened in a vertical split, the filter buffer will be opened
below it.